requirements.txt
selenium~=4.6.0
google-cloud-speech~=2.16.2
requests~=2.28.1
